Refactor gauge colors into redux state instead of CEO state
